recipe_community/requirements.txt
asgiref==3.3.1
Django==3.1.7
django-filter==2.4.0
flake8==3.8.4
mccabe==0.6.1
Pillow==8.1.2
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
python-dotenv==0.15.0
pytz==2021.1
reportlab==3.5.66
sorl-thumbnail==12.7.0
sqlparse==0.4.1
